split login call to api into its own module
